YY -- "sighting" now works. we will just "trust" that the clients send sensible "projectile fired" messages to the server; how the client figures that out is the client's problem.
