+++ client/src/lib.rs
@@ -1,3 +1,4 @@
+use bevy_rapier3d::prelude::{QueryFilter, RapierContext};
 use bevy_renet::{
     renet::{ClientAuthentication, DefaultChannel, RenetClient, RenetConnectionConfig},
     run_if_client_connected, RenetClientPlugin,
@@ -329,5 +330,63 @@ pub fn visualize_projectiles(
                     ..default()
                 });
             });
+    }
+}
+
+pub fn handle_projectile_engagement(
+    mass_query: Query<
+        (&Transform, &components::MassID),
+        (
+            Without<components::Inhabitable>,
+            Without<components::ClientInhabited>,
+        ),
+    >,
+    mut sights_visibility: Query<&mut Visibility, With<components::Sights>>,
+    inhabited_mass_query: Query<
+        (&Transform, &components::MassID),
+        With<components::ClientInhabited>,
+    >,
+    rapier_context: Res<RapierContext>,
+    keys: Res<Input<KeyCode>>,
+    mut to_server_events: EventWriter<events::ToServer>,
+) {
+    if inhabited_mass_query.is_empty() {
+        error!("No `ClientInhabited` masses found!");
+    }
+    if let Ok((client_pov, &components::MassID(from_mass_id))) = inhabited_mass_query.get_single() {
+        let ray_origin = client_pov.translation;
+        let ray_direction = -client_pov.local_z();
+        let intersection = rapier_context.cast_ray(
+            ray_origin,
+            ray_direction,
+            150.0,
+            false,
+            QueryFilter::only_dynamic(),
+        );
+        if let Some((mass, distance)) = intersection {
+            if let Ok((mass_transform, &components::MassID(to_mass_id))) = mass_query.get(mass) {
+                sights_visibility.for_each_mut(|mut visibility| visibility.is_visible = true);
+                debug!("Mass {to_mass_id} is now in our sights");
+                if keys.just_pressed(KeyCode::Space) {
+                    let global_impact_site = ray_origin + (ray_direction * distance);
+                    let local_impact_direction =
+                        (global_impact_site - mass_transform.translation).normalize();
+                    let launch_time = SystemTime::now()
+                        .duration_since(SystemTime::UNIX_EPOCH)
+                        .unwrap()
+                        .as_millis();
+                    to_server_events.send(events::ToServer::ProjectileFired(
+                        events::ProjectileFlight {
+                            launch_time,
+                            from_mass_id,
+                            to_mass_id,
+                            local_impact_direction,
+                        },
+                    ));
+                }
+            }
+        } else {
+            sights_visibility.for_each_mut(|mut visibility| visibility.is_visible = false);
+        }
     }
 }
